+++ modules/Runes.lua
@@ -185,7 +185,7 @@ function Runes.prototype:GetOptions()
 		type = 'range',
 		name = L["Rune gap"],
 		desc = L["Spacing between each rune (only works for graphical mode)"],
-		min = 0,
+		min = -20,
 		max = 100,
 		step = 1,
 		get = function()